+++ src/TemperatureDataBaseWPF/Services/Service.cs
@@ -1,12 +1,14 @@
-﻿namespace TemperatureDataBaseWPF.Services
+namespace TemperatureDataBaseWPF.Services
 {
     public class Service : IService
     {
         private SerialDataAcquisition _serialHandler;
+        private LoadController _loadController;
 
         public Service()
         {
             _serialHandler = new SerialDataAcquisition();
+            _loadController = new LoadController();
         }
 
         #region Interface